move alertPrimaryButton
