+++ app/src/atoms/buttons/index.ts
@@ -1,6 +1,5 @@
 export { TertiaryButton } from './TertiaryButton'
 export { QuaternaryButton } from './QuaternaryButton'
-export { AlertPrimaryButton } from './AlertPrimaryButton'
 export { SubmitPrimaryButton } from './SubmitPrimaryButton'
 export { ToggleButton } from './ToggleButton'
 export { BackButton } from './BackButton'
