+++ app/src/organisms/Devices/RobotSettings/AdvancedTab/AdvancedTabSlideouts/FactoryResetModal.tsx
@@ -8,6 +8,7 @@ import {
   Flex,
   DIRECTION_COLUMN,
   JUSTIFY_FLEX_END,
+  AlertPrimaryButton,
   SPACING,
   Link,
   TYPOGRAPHY,
@@ -15,7 +16,6 @@ import {
   PrimaryButton,
 } from '@opentrons/components'
 import { StyledText } from '../../../../../atoms/text'
-import { AlertPrimaryButton } from '../../../../../atoms/buttons'
 import { Modal } from '../../../../../molecules/Modal'
 import {
   useDispatchApiRequest,
