+++ app/src/organisms/Devices/RobotSettings/ConnectNetwork/DisconnectModal.tsx
@@ -13,10 +13,10 @@ import {
   SPACING,
   TYPOGRAPHY,
   PrimaryButton,
+  AlertPrimaryButton,
   JUSTIFY_CENTER,
 } from '@opentrons/components'
 
-import { AlertPrimaryButton } from '../../../../atoms/buttons'
 import { StyledText } from '../../../../atoms/text'
 import { Modal } from '../../../../molecules/Modal'
 import { useRobot } from '../../../../organisms/Devices/hooks'
