+++ app/src/organisms/LabwarePositionCheck/ExitConfirmation.tsx
@@ -8,13 +8,13 @@ import {
   DIRECTION_COLUMN,
   SPACING,
   SIZE_3,
+  AlertPrimaryButton,
   JUSTIFY_CENTER,
   COLORS,
   TYPOGRAPHY,
   SecondaryButton,
 } from '@opentrons/components'
 import { StyledText } from '../../atoms/text'
-import { AlertPrimaryButton } from '../../atoms/buttons'
 import { NeedHelpLink } from '../CalibrationPanels'
 
 const LPC_HELP_LINK_URL =
