+++ components/src/atoms/index.tsx
@@ -1,3 +1,4 @@
 export { PrimaryButton } from './buttons/PrimaryButton'
 export { SecondaryButton } from './buttons/SecondaryButton'
+export { AlertPrimaryButton } from './buttons/AlertPrimaryButton'
 export { CheckboxField } from './CheckboxField'
